Remove Clinical Mongo Connection
